added hover handlers for card in "Other places in the neighbourhood" section
