+++ src/hooks/index.ts
@@ -1,9 +1,11 @@
 import useMap from './use-map.ts';
 import useSetCityLocation from './use-set-city-location.ts';
 import useSetMarkers from './use-set-markers.ts';
+import useHoverHandlers from './use-hover-handlers.ts';
 
 export {
   useMap,
   useSetCityLocation,
   useSetMarkers,
+  useHoverHandlers,
 };
